+++ create-vtj/src/options.ts
@@ -36,6 +36,11 @@ export const templates: ITemplate[] = [
     display: '物料开发项目( material )',
     color: lightRed
   },
+  {
+    name: 'extension',
+    display: '设计器扩展开发项目( extension )',
+    color: lightRed
+  },
   // {
   //   name: 'nest',
   //   display: '服务端应用( nest )',