+++ platforms/pro/src/Auth.vue
@@ -3,27 +3,36 @@
 </template>
 <script lang="ts" setup>
   import { useRoute } from 'vue-router';
-  import { createAdapter, type CreateAdapterOptions } from '@vtj/pro';
+  import { createAccess } from '@vtj/pro';
   import { ElMessageBox } from 'element-plus';
+  import { jsonp } from '@vtj/utils';
 
   export interface Props {
-    options?: CreateAdapterOptions;
+    remote?: string;
+    access?: Record<string, any>;
   }
 
-  const props = defineProps<Props>();
-
-  const _options = {
-    ...props.options,
-    access: {
-      alert: ElMessageBox.alert,
-      ...props.options?.access
+  const props = withDefaults(defineProps<Props>(), {
+    remote: 'https://lcdp.vtj.top',
+    access: () => {
+      return {
+        auth: 'https://lcdp.vtj.top/auth.html',
+        storageKey: 'RRO_IDE_ACCESS_STORAGE__',
+        privateKey:
+          'MIIBOgIBAAJBAKoIzmn1FYQ1YOhOBw9EhABxZ+PySAIaydI+zdhoKflrdgJ4A5E4/5gbQmRpk09hPWG8nvX7h+l/QLU8kXxAIBECAwEAAQJAAlgpxQY6sByLsXqzJcthC8LSGsLf2JEJkHwlnpwFqlEV8UCkoINpuZ2Wzl+aftURu5rIfAzRCQBvHmeOTW9/zQIhAO5ufWDmnSLyfAAsNo5JRNpVuLFCFodR8Xm+ulDlosR/AiEAtpAltyP9wmCABKG/v/hrtTr3mcvFNGCjoGa9bUAok28CIHbrVs9w1ijrBlvTsXYwJw46uP539uKRRT4ymZzlm9QjAiB+1KH/G9f9pEEL9rtaSOG7JF5D0JcOjlze4MGVFs+ZrQIhALKOUFBNr2zEsyJIjw2PlvEucdlG77UniszjXTROHSPd'
+      };
     }
-  };
+  });
+
+  const access = createAccess({
+    alert: ElMessageBox.alert,
+    ...props.access
+  });
+
   const route = useRoute();
-  const adapter = createAdapter(_options);
   const getLoginInfo = async (token: string) => {
-    const api = `${adapter.remote}/api/open/user/${token}`;
-    const res = await adapter.jsonp(api).catch(() => null);
+    const api = `${props.remote}/api/open/user/${token}`;
+    const res = await jsonp(api).catch(() => null);
     if (!res || !res.success) {
       await ElMessageBox.alert('登录失败');
       return null;
@@ -36,7 +45,7 @@
   const info = token ? await getLoginInfo(token) : null;
   if (info) {
     try {
-      adapter.access?.login(info);
+      access.login(info);
       const redirect = route.query.redirect as string;
       location.href = decodeURIComponent(redirect || '/');
     } catch (e) {
@@ -44,3 +53,9 @@
     }
   }
 </script>
+
+<style lang="scss" scoped>
+  .loading {
+    padding: 40px;
+  }
+</style>