+++ platforms/pro/src/extension.ts
@@ -6,50 +6,43 @@ import * as designer from '@vtj/designer';
 import * as renderer from '@vtj/renderer';
 import * as VtjIcons from '@vtj/icons';
 import * as ElementPlus from 'element-plus';
-import type { ExtensionConfig } from '@vtj/core';
+import type { VTJConfig } from '@vtj/core';
 import type { EngineOptions } from '@vtj/designer';
 
-export type ExtensionOptions = ExtensionConfig;
-export type ExtensionFactory = (
-  config: ExtensionConfig
-) => Partial<EngineOptions> | void;
+export type ExtensionFactory =
+  | Partial<EngineOptions>
+  | ((config: VTJConfig, ...args: any) => Partial<EngineOptions>);
 
-export interface ExtensionOutput {
-  options: Partial<EngineOptions>;
-  adapters: Record<string, any>;
-}
+export type ExtensionOutput = Partial<EngineOptions>;
 
 export class Extension {
   private urls: string[] = [];
   private library: string = '';
   private params: any[] = [];
   private __BASE_PATH__: string = '/';
-  private __adapters__: Record<string, any> = {};
-  constructor(private options: ExtensionOptions) {
+  constructor(private config: VTJConfig) {
     const __VTJ_PRO__ = {
       ...core,
       ...designer,
       ...renderer
     };
-
-    (window as any).Vue = Vue;
-    (window as any).__VTJ_PRO__ = __VTJ_PRO__;
-    (window as any).VtjUtils = VtjUtils;
-    (window as any).VtjIcons = VtjIcons;
-    (window as any).VtjUI = VtjUI;
-    (window as any).ElementPlus = ElementPlus;
-    const {
-      urls = [],
-      library,
-      params = [],
-      __BASE_PATH__ = '/',
-      __adapters__ = {}
-    } = options || {};
+    const globals: Record<string, any> = {
+      Vue,
+      __VTJ_PRO__,
+      VtjUtils,
+      VtjIcons,
+      VtjUI,
+      ElementPlus
+    };
+    for (const [name, value] of Object.entries(globals)) {
+      (window as any)[name] = value;
+    }
+    const { extension, __BASE_PATH__ = '/' } = config || {};
+    const { urls = [], library, params = [] } = extension || {};
     this.urls = urls;
     this.library = library;
     this.params = params;
     this.__BASE_PATH__ = __BASE_PATH__;
-    this.__adapters__ = __adapters__;
   }
   async load(): Promise<ExtensionOutput> {
     let options: Partial<EngineOptions> = {};
@@ -63,19 +56,16 @@ export class Extension {
         .map((n) => `${base}${n}`);
       renderer.loadCssUrl(css);
       if (scripts.length) {
-        const output = await renderer
+        const output: ExtensionFactory = await renderer
           .loadScriptUrl(scripts, this.library)
           .catch(() => null);
         if (output && typeof output === 'function') {
-          options = output.call(output, this.options, this.params);
+          options = output.call(output, this.config, ...this.params);
         } else {
           options = output || {};
         }
       }
     }
-    return {
-      options,
-      adapters: this.__adapters__
-    };
+    return options;
   }
 }