+++ apps/material/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.1](https://gitee.com/newgateway/vtj/compare/vtj-material@0.11.1-alpha.1...vtj-material@0.11.1) (2025-03-18)
+
+**Note:** Version bump only for package vtj-material
+
+
+
+
+
+## [0.11.1-alpha.1](https://gitee.com/newgateway/vtj/compare/vtj-material@0.2.1-alpha.0...vtj-material@0.11.1-alpha.1) (2025-03-18)
+
+**Note:** Version bump only for package vtj-material
+
+
+
+
+
+## [0.2.1-alpha.0](https://gitee.com/newgateway/vtj/compare/vtj-material@0.2.0...vtj-material@0.2.1-alpha.0) (2025-03-18)
+
+**Note:** Version bump only for package vtj-material
+
+
+
+
+
+# [0.2.0](https://gitee.com/newgateway/vtj/compare/vtj-material@0.1.3...vtj-material@0.2.0) (2025-03-17)
+
+**Note:** Version bump only for package vtj-material
+
+
+
+
+
 ## [0.1.3](https://gitee.com/newgateway/vtj/compare/vtj-material@0.1.2...vtj-material@0.1.3) (2025-03-14)
 
 **Note:** Version bump only for package vtj-material